Remove JSON and use library
Also remove bundling and instead rely on locally file.
