+++ package.json
@@ -30,10 +30,17 @@
   "homepage": "https://github.com/nearprotocol/near-bindgen#readme",
   "devDependencies": {
     "assemblyscript": "^0.8.1",
+    "assemblyscript-json": "nearprotocol/assemblyscript-json",
     "near-runtime-ts": "^0.5.1",
     "ts-loader": "^6.2.0",
+    "ts-node": "^8.6.2",
     "typescript": "^3.6.4",
     "webpack": "^4.41.0",
     "webpack-cli": "^3.3.9"
-  }
+  },
+  "files": [
+    "index.js",
+    "assembly/index.ts",
+    "compiler.js"
+  ]
 }
