+++ webpack.config.js
@@ -6,4 +6,4 @@ const files = {
 
 const outfolder = path.resolve(__dirname);
 
-module.exports = config(files, outfolder,  ["./assembly/nearEntry.ts"]);
+module.exports = config(files, outfolder);
